chore - move auto-approval check tests to correct folder
