+++ test/unit/services/auto-approval/checks/test-is-prison-not-in-guernsey-jersey.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect
-const guernseyJerseyPrisonsEnum = require('../../../../app/constants/guernsey-jersey-prisons-enum')
-const isPrisonNotInGuernseyJersey = require('../../../../app/services/auto-approval/checks/is-prison-not-in-guernsey-jersey')
+const guernseyJerseyPrisonsEnum = require('../../../../../app/constants/guernsey-jersey-prisons-enum')
+const isPrisonNotInGuernseyJersey = require('../../../../../app/services/auto-approval/checks/is-prison-not-in-guernsey-jersey')
 
 var validAutoApprovalData = {
   Prisoner: {
